libtk/pool/freetype: route benign int<->void* casts through uintptr (LP64 hygiene)
These sites stash a small int (a handle/id/fd, never a real pointer) in a void*
slot and read it back -- safe round-trips, but gcc's -Wpointer-to-int-cast /
-Wint-to-pointer-cast can't tell them apart from a genuine pointer truncation,
so ~40 benign warnings were burying any real one.  Bridge each through uintptr
(pointer-width) so the conversion is deliberate and warning-free; a real
pointer->int truncation introduced later now stands out as the only such warning.

- include/tk.h: AUXI/IAUX go through uintptr (+ contract comment); covers the
  libtk users (parse.c).
- include/pool.h: BHDRSIZE offsetof idiom via uintptr (clears the alloc.c hits).
- libtk/{utils,parse,ebind,buton}.c: open-coded int-cookie casts -> AUXI/IAUX.
- libfreetype/ftsystem_inf.c: STREAM_FD / descriptor.pointer (int fd) via uintptr.

Deliberately NOT touched: the JIT jump-table / PC casts in libinterp/xec.c and
emu/port/{devprog,devprof}.c.  Those truncate a native address to 32 bits ON
PURPOSE (the code buffer is MAP_32BIT and the value is read back *sign*-extended,
see comp-amd64.c); bridging them through unsigned uintptr would zero-extend and
break that contract.  They remain the only pointer/int-cast warnings, by design.
ON_JIT.md (low-2GB arena bullet) now records this so a future warning-cleanup
pass doesn't fall into that trap.

+++ include/tk.h
@@ -135,8 +135,13 @@ enum
 #define TKI2F(i)	((i)*Tkfpscalar)
 extern	int TKF2I(int);
 /*#define TKF2I(f)	(((f) + Tkfpscalar/2)/Tkfpscalar)*/
-#define IAUX(i)		((void*)i)
-#define AUXI(i)		((int)i)
+/* Stash a small int (a handle/id, never a real pointer) in a void* slot and
+ * recover it, width-safe on ILP32 and LP64.  The uintptr bridge makes the
+ * pointer<->int conversion deliberate, so it neither warns nor truncates a
+ * value that genuinely fits in an int.  Contract: the slot only ever holds
+ * such an int -- never a host pointer (which AUXI WOULD truncate). */
+#define IAUX(i)		((void*)(uintptr)(i))
+#define AUXI(i)		((int)(uintptr)(i))
 #define TKKEY(i)	((i)&0xFFFF)
 
 typedef struct Tk Tk;
